Fix most MSVC compiler warnings
# Details

- Rename `NO_DEFAULT_GCC()` to `NO_DEFAULT()` and add its MSVC branch (`__assume(0)`), fixing C4715 in 27 functions.
- Add `/wd4200` for Windows since flexible array members cannot be declared otherwise (164 warnings).
- Replace the `(-size) & 3` padding idiom with `Utils::Byte::PadTo4Bytes()` in RTCP, fixing C4146 (31 warnings)
- Cast the `size_t` returned by `RTC::SCTP::InterleavedReassemblyStreams::Stream::TryToAssembleMessages()` before negating it
- Give `RTC::SCTP::ReassemblyQueue::HighWatermarkLimit` a float literal suffix, fixing C4305

+++ worker/src/RTC/SCTP/rx/InterleavedReassemblyStreams.cpp
@@ -118,7 +118,7 @@ namespace RTC
 				this->nextMid.Increment();
 
 				// This might unblock assembling more messages.
-				return -TryToAssembleMessages();
+				return -static_cast<int32_t>(TryToAssembleMessages());
 			}
 
 			// Slow path.
